+++ osu.Game/Screens/Edit/Screens/Compose/Compose.cs
@@ -8,7 +8,6 @@ using osu.Framework.Graphics.Containers;
 using osu.Framework.Graphics.Shapes;
 using osu.Framework.Logging;
 using osu.Game.Beatmaps;
-using osu.Game.Screens.Edit.Screens.Compose.BeatSnap;
 using osu.Game.Screens.Edit.Screens.Compose.Timeline;
 
 namespace osu.Game.Screens.Edit.Screens.Compose
