+++ osu.Game/osu.Game.csproj
@@ -380,7 +380,7 @@
     <Compile Include="Rulesets\Replays\Types\IConvertibleReplayFrame.cs" />
     <Compile Include="Rulesets\Scoring\Legacy\LegacyScoreParser.cs" />
     <Compile Include="Rulesets\UI\JudgementContainer.cs" />
-    <Compile Include="Screens\Edit\Screens\Compose\BeatSnap\BeatSnapVisualiser.cs" />
+    <Compile Include="Screens\Edit\Screens\Compose\DrawableBeatDivisor.cs" />
     <Compile Include="Screens\Edit\Screens\Compose\Layers\BorderLayer.cs" />
     <Compile Include="Screens\Edit\Screens\Compose\Layers\HitObjectMaskLayer.cs" />
     <Compile Include="Screens\Edit\Screens\Compose\Layers\SelectionBox.cs" />
